Upload handling changed - tests updated
