+++ resources/images/controller/images_external_test.go
@@ -24,7 +24,6 @@ import (
 	"net/http"
 	"net/textproto"
 	"os"
-	"path"
 	"testing"
 	"time"
 
@@ -40,9 +39,6 @@ import (
 	h "github.com/mendersoftware/deployments/utils/testing"
 	"github.com/mendersoftware/go-lib-micro/requestid"
 	"github.com/mendersoftware/go-lib-micro/requestlog"
-	"github.com/mendersoftware/mender-artifact/parser"
-	atutils "github.com/mendersoftware/mender-artifact/test_utils"
-	"github.com/mendersoftware/mender-artifact/writer"
 	"github.com/satori/go.uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -51,6 +47,8 @@ import (
 // Notice: 	Controller tests are not pure unit tests,
 // 			they are more of integration test beween controller and view
 //			testing actuall HTTP endpoint input/reponse
+
+const validUUIDv4 = "d50eda0d-2cea-4de1-8d42-9cd3e7e8670d"
 
 type fakeImageModeler struct {
 	getImage          *images.SoftwareImage
@@ -88,9 +86,8 @@ func (fim *fakeImageModeler) DeleteImage(imageID string) error {
 }
 
 func (fim *fakeImageModeler) CreateImage(
-	artifact io.Reader,
 	metaConstructor *images.SoftwareImageMetaConstructor,
-	metaArtifactConstructor *images.SoftwareImageMetaArtifactConstructor) (string, error) {
+	artifact io.Reader) (string, error) {
 	return "", nil
 }
 
@@ -140,7 +137,7 @@ func TestControllerGetImage(t *testing.T) {
 	// have image, get OK
 	imageMeta := images.NewSoftwareImageMetaConstructor()
 	imageMetaArtifact := images.NewSoftwareImageMetaArtifactConstructor()
-	constructorImage := images.NewSoftwareImage(imageMeta, imageMetaArtifact)
+	constructorImage := images.NewSoftwareImage(validUUIDv4, imageMeta, imageMetaArtifact)
 	imagesModel.getImageError = nil
 	imagesModel.getImage = constructorImage
 	recorded = test.RunRequest(t, api.MakeHandler(),
@@ -170,7 +167,7 @@ func TestControllerListImages(t *testing.T) {
 	imagesModel.listImagesError = nil
 	imageMeta := images.NewSoftwareImageMetaConstructor()
 	imageMetaArtifact := images.NewSoftwareImageMetaArtifactConstructor()
-	constructorImage := images.NewSoftwareImage(imageMeta, imageMetaArtifact)
+	constructorImage := images.NewSoftwareImage(validUUIDv4, imageMeta, imageMetaArtifact)
 	imagesModel.imagesList = append(imagesModel.imagesList, constructorImage)
 	recorded = test.RunRequest(t, api.MakeHandler(),
 		test.MakeSimpleRequest("GET", "http://localhost/api/0.0.1/images", nil))
@@ -186,7 +183,7 @@ func TestControllerDeleteImage(t *testing.T) {
 
 	imageMeta := images.NewSoftwareImageMetaConstructor()
 	imageMetaArtifact := images.NewSoftwareImageMetaArtifactConstructor()
-	constructorImage := images.NewSoftwareImage(imageMeta, imageMetaArtifact)
+	constructorImage := images.NewSoftwareImage(validUUIDv4, imageMeta, imageMetaArtifact)
 
 	// wrong id
 	recorded := test.RunRequest(t, api.MakeHandler(),
@@ -262,15 +259,12 @@ func TestControllerEditImage(t *testing.T) {
 func TestSoftwareImagesControllerNewImage(t *testing.T) {
 	t.Parallel()
 
-	// create temp dir
-	td, _ := ioutil.TempDir("", "mender-install-update-")
-	defer os.RemoveAll(td)
-	// make a fake update artifact
-	upath, err := makeFakeUpdate(t, path.Join(td, "update-root"), true)
-	// open archive file
-	imageBody, err := ioutil.ReadFile(upath)
+	file := createValidImageFile()
+	imageBody, err := ioutil.ReadAll(file)
 	assert.NoError(t, err)
 	assert.NotNil(t, imageBody)
+	defer os.Remove(file.Name())
+	defer file.Close()
 
 	testCases := []struct {
 		h.JSONResponseParams
@@ -306,7 +300,7 @@ func TestSoftwareImagesControllerNewImage(t *testing.T) {
 			InputContentType: "multipart/form-data",
 			JSONResponseParams: h.JSONResponseParams{
 				OutputStatus:     http.StatusBadRequest,
-				OutputBodyObject: h.ErrorToErrStruct(errors.New("Validating metadata: Name: non zero value required;")),
+				OutputBodyObject: h.ErrorToErrStruct(errors.New("Name: non zero value required;")),
 			},
 		},
 		{
@@ -320,7 +314,7 @@ func TestSoftwareImagesControllerNewImage(t *testing.T) {
 			InputContentType: "multipart/form-data",
 			JSONResponseParams: h.JSONResponseParams{
 				OutputStatus:     http.StatusBadRequest,
-				OutputBodyObject: h.ErrorToErrStruct(errors.New("Validating metadata: Name: non zero value required;")),
+				OutputBodyObject: h.ErrorToErrStruct(errors.New("Name: non zero value required;")),
 			},
 		},
 		{
@@ -366,34 +360,7 @@ func TestSoftwareImagesControllerNewImage(t *testing.T) {
 			InputContentType: "multipart/form-data",
 			JSONResponseParams: h.JSONResponseParams{
 				OutputStatus:     http.StatusBadRequest,
-				OutputBodyObject: h.ErrorToErrStruct(errors.New("Last part should be an image")),
-			},
-		},
-		{
-			InputBodyObject: []Part{
-				Part{
-					FieldName:  "name",
-					FieldValue: "n",
-				},
-				Part{
-					FieldName:  "device_type",
-					FieldValue: "dt",
-				},
-				Part{
-					FieldName:  "yocto_id",
-					FieldValue: "yi",
-				},
-				Part{
-					FieldName:   "artifact",
-					ContentType: "application/octet-stream",
-					ImageData:   []byte{0},
-				},
-			},
-			InputContentType: "multipart/form-data",
-			InputModelID:     "1234",
-			JSONResponseParams: h.JSONResponseParams{
-				OutputStatus:     http.StatusBadRequest,
-				OutputBodyObject: h.ErrorToErrStruct(errors.New("reading artifact error: reader: error reading archive: unexpected EOF")),
+				OutputBodyObject: h.ErrorToErrStruct(errors.New("The last part of the multipart/form-data message should be an image.")),
 			},
 		},
 		{
@@ -464,12 +431,11 @@ func TestSoftwareImagesControllerNewImage(t *testing.T) {
 
 		model.On(
 			"CreateImage",
+			mock.AnythingOfType("*images.SoftwareImageMetaConstructor"),
 			mock.MatchedBy(func(ir interface{}) bool {
 				_, ok := ir.(io.Reader)
 				return ok
-			}),
-			mock.AnythingOfType("*images.SoftwareImageMetaConstructor"),
-			mock.AnythingOfType("*images.SoftwareImageMetaArtifactConstructor")).
+			})).
 			Return(testCase.InputModelID, testCase.InputModelError)
 
 		api := setUpRestTest("/r", rest.Post, NewSoftwareImagesController(model, new(view.RESTView)).NewImage)
@@ -519,36 +485,11 @@ func MakeMultipartRequest(method string, urlStr string, contentType string, payl
 	return r
 }
 
-func makeFakeUpdate(t *testing.T, root string, valid bool) (string, error) {
-
-	var dirStructOK = []atutils.TestDirEntry{
-		{Path: "0000", IsDir: true},
-		{Path: "0000/data", IsDir: true},
-		{Path: "0000/data/update.ext4", Content: []byte("first update"), IsDir: false},
-		{Path: "0000/type-info", Content: []byte(`{"type": "rootfs-image"}`), IsDir: false},
-		{Path: "0000/meta-data", Content: []byte(`{"DeviceType": "vexpress-qemu", "ImageID": "core-image-minimal-201608110900"}`), IsDir: false},
-		{Path: "0000/signatures", IsDir: true},
-		{Path: "0000/signatures/update.sig", IsDir: false},
-		{Path: "0000/scripts", IsDir: true},
-		{Path: "0000/scripts/pre", IsDir: true},
-		{Path: "0000/scripts/pre/0000_install.sh", Content: []byte("run me!"), IsDir: false},
-		{Path: "0000/scripts/post", IsDir: true},
-		{Path: "0000/scripts/check", IsDir: true},
-	}
-
-	err := atutils.MakeFakeUpdateDir(root, dirStructOK)
-	assert.NoError(t, err)
-
-	aw := awriter.NewWriter("mender", 1, []string{"vexpress"}, "mender-1.0")
-
-	rp := &parser.RootfsParser{}
-	aw.Register(rp)
-
-	upath := path.Join(root, "update.tar")
-	err = aw.Write(root, upath)
-	assert.NoError(t, err)
-
-	return upath, nil
+func createValidImageFile() *os.File {
+	someData := []byte{115, 111, 109, 101, 10, 11}
+	tmpfile, _ := ioutil.TempFile("", "artifact-")
+	tmpfile.Write(someData)
+	return tmpfile
 }
 
 func TestSoftwareImagesControllerDownloadLink(t *testing.T) {
